pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "job-bot"
version = "0.1.0"
description = "Automated job scraper and applicator using Claude AI"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "playwright>=1.44.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "rich>=13.0.0",
    "tenacity>=8.0.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
]

[project.scripts]
job-bot = "cli:app"

[tool.hatch.build.targets.wheel]
packages = ["job_bot"]

[tool.ruff]
line-length = 100
target-version = "py312"
